requirements.txt
opencv-python
mediapipe
ultralytics
numpy
python-dotenv
mss
Pillow
pymongo
lap